reorganizing the apply feasible trajectory operation to be available for all phased arrays and for all types of trajectories
